Arm changes and Controller lockouts
setRaw1 and setRaw2 made to be able to lockout on of them when buttons are pressed.
ArmBasetoTop() moves arm to the top at a accelerating and decelerating rate.
button A and button B begin lift and move arm to top respectively.
when A or B is in action Controller 2 is locked out.